+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Change Log
 0.1.11 (2014-09-01*)
 --------------------
 1. Refined router and middleware code tpl;
+2. Added nodemon into env required listing;
 
 In progress...